+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - **Breaking**: The `--folder` argument to the `occ memories:index` has been changed to `--path`
+- **Feature**: Allow right-click on image in viewer ([#1479](https://github.com/pulsejet/memories/issues/1479), [#993](https://github.com/pulsejet/memories/issues/993), [#1398](https://github.com/pulsejet/memories/issues/1398))
 
 ## [v7.6.0] - 2025-07-03
 
-- **Feature**: Looping of live photos is now gated behind a user setting ([[#1457](https://github.com/pulsejet/memories/issues/1457)])
+- **Feature**: Looping of live photos is now gated behind a user setting ([#1457](https://github.com/pulsejet/memories/issues/1457))
 - **Feature**: Option to hide sharing user from timeline ([#1452](https://github.com/pulsejet/memories/issues/1452))
 - **Fix**: Hide sharing user from public links ([#1452](https://github.com/pulsejet/memories/issues/1452))
 - **Fix**: Broken Google Takeout migration